Move over RPY matrix tools, and add to depth control
