+++ thruster_control/src/thruster_control/geometry_helper.py
@@ -3,6 +3,8 @@ import numpy as np
 
 from geometry_msgs.msg import Wrench
 
+import aquadrone_math_utils.orientation_math as OH
+
 def vector_to_wrench(v):
     w = Wrench()
     w.force.x = v[0]
@@ -25,7 +27,7 @@ def get_thruster_wrench_vector(x, y, z, roll, pitch, yaw):
     return np.vstack((force, moment))
 
 def get_thruster_force(r, p, y):
-    return np.dot(RPY_Matrix(r,p,y), thruster_force_dir_vec())
+    return np.dot(OH.RPY_Matrix(r,p,y), thruster_force_dir_vec())
 
 def thruster_force_dir_vec():
     # Constant direction that a thruster applies force in
@@ -34,31 +36,3 @@ def thruster_force_dir_vec():
     v.shape = (3,1)
     return v
 
-def Yaw(y):
-        return np.matrix([
-                [math.cos(y), -math.sin(y), 0],
-                [math.sin(y),  math.cos(y), 0],
-                [          0,            0, 1]
-                ])
-
-def Pitch(p):
-        return np.matrix([
-                [ math.cos(p), 0, math.sin(p)],
-                [               0, 1,               0],
-                [-math.sin(p), 0, math.cos(p)]
-                ])
-
-def Roll(r):
-        return np.matrix([
-                [1,              0,               0],
-                [0, math.cos(r), -math.sin(r)],
-                [0, math.sin(r),  math.cos(r)]
-                ])
-
-def RPY_Matrix(r, p, y):
-    r = Roll(r)
-    p = Pitch(p)
-    y = Yaw(y)
-
-    interm = np.dot(y,p)
-    return np.dot(interm, r)
